# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Price {
  id: string;
  label: string;
  slides: string;
  usd: number;
  inr: number;
  gel: number;
}

export interface SubjectGroup {
  id: string;
  label: string;
  description?: string;
  subjects: string[];
  prices?: Price[];
}

export interface Catalog {
  groups: SubjectGroup[];
  currencies: string[];
}

export type OrderInputDeliveryMode = typeof OrderInputDeliveryMode[keyof typeof OrderInputDeliveryMode];


export const OrderInputDeliveryMode = {
  standard: 'standard',
  personalized: 'personalized',
} as const;

export type OrderInputCurrency = typeof OrderInputCurrency[keyof typeof OrderInputCurrency];


export const OrderInputCurrency = {
  USD: 'USD',
  INR: 'INR',
  GEL: 'GEL',
} as const;

export interface OrderInput {
  email: string;
  /** @minLength 1 */
  subjectGroup: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  slideRange: string;
  /** @minLength 3 */
  topic: string;
  instructions?: string;
  deliveryMode: OrderInputDeliveryMode;
  currency?: OrderInputCurrency;
}

export interface Order {
  id: string;
  status: string;
  email: string;
  topic: string;
  deliveryEta: string;
  /** @nullable */
  checkoutUrl?: string | null;
}

export type CheckoutInputCurrency = typeof CheckoutInputCurrency[keyof typeof CheckoutInputCurrency];


export const CheckoutInputCurrency = {
  USD: 'USD',
  INR: 'INR',
  GEL: 'GEL',
} as const;

export interface CheckoutInput {
  orderId: string;
  currency: CheckoutInputCurrency;
}

export interface Checkout {
  purchaseUrl: string;
  orderId?: string;
}

export interface ErrorResponse {
  error: string;
}