# lib/api-zod/src/generated/types/catalog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectGroup } from './subjectGroup';

export interface Catalog {
  groups: SubjectGroup[];
  currencies: string[];
}