# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutInputCurrency } from './checkoutInputCurrency';

export interface CheckoutInput {
  orderId: string;
  currency: CheckoutInputCurrency;
}