# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: string;
  status: string;
  email: string;
  topic: string;
  deliveryEta: string;
  /** @nullable */
  checkoutUrl?: string | null;
}