# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputCurrency } from './orderInputCurrency';
import type { OrderInputDeliveryMode } from './orderInputDeliveryMode';

export interface OrderInput {
  email: string;
  /** @minLength 1 */
  subjectGroup: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  slideRange: string;
  /** @minLength 3 */
  topic: string;
  instructions?: string;
  deliveryMode: OrderInputDeliveryMode;
  currency?: OrderInputCurrency;
}