# lib/api-zod/src/generated/types/price.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Price {
  id: string;
  label: string;
  slides: string;
  usd: number;
  inr: number;
  gel: number;
}