# lib/api-zod/src/generated/types/subjectGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Price } from './price';

export interface SubjectGroup {
  id: string;
  label: string;
  description?: string;
  subjects: string[];
  prices?: Price[];
}